requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
#google-generativeai==0.4.1
orjson==3.10.0
azure-functions==1.18.0
nest-asyncio==1.6.0
mysql-connector-python==8.3.0
Jinja2==3.1.3
pypdf==4.0.0
PyPDF2==3.0.1
pandas==2.2.0
chromadb==0.5.0
google-generativeai==0.5.2
pysqlite3-binary==0.5.2.post3
openai
langchain==0.1.14
langchain-core
langchain-community
tiktoken
langchain_google_genai
langchain_openai